sha1dc: remove in favor of using sha1collisiondetection as a submodule
Remove the copy of the sha1collisiondetection's submodule in sha1dc,
making the submodule a hard requirement. It was the default before,
but now it's a hard dependency.

The main benefit of doing this is to get the git project itself to
dogfood submodules & have the entire community enjoy the resulting
fixes that'll come out of that. Not that it's a big bother for us to
maintain the sha1dc/ & sha1collisiondetection/ directories and we need
to have a submodule for our own use. It also saves us the work of
updating both copies of the code if and when upstream changes.

There is still a check for NO_DC_SHA1_SUBMODULE=NoThanks which'll
indicate that the submodule is now a hard dependency. That check
should not fire if DC_SHA1_EXTERNAL=YesPlease is provide, as packagers
may have supplied both to link to the external library, making it
error out in that case would be a bother.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>