sha1dc: remove in favor of using sha1collisiondetection as a submodule
Follow-up the preceding changes to require
NO_DC_SHA1_SUBMODULE=UnfortunatelyYes to use the sha1dc/ code by
completely removing it. This ends the saga of getting a submodule into
git.git as a hard dependency, and saves us from updating both copies
of the code if and when upstream changes.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>